Baseline metric calculation task created to calculate Flame Length and Rate of Spread variables
